Under refactor of separate type files, we can do that later across all resources if we want
