+++ workos/types/events/directory_user_with_previous_attributes.py
@@ -1,4 +1,4 @@
-from workos.types.directory_sync.directory_user import DirectoryUser
+from workos.resources.directory_sync import DirectoryUser
 from workos.types.events.previous_attributes import PreviousAttributes
